social_media_agent/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asn1crypto==1.5.1
attrs==25.1.0
bip-utils==2.9.3
bitarray==3.0.0
cachetools==5.5.1
cbor2==5.6.5
cdp-agentkit-core==0.0.11
cdp-langchain==0.0.13
cdp-sdk==0.15.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
ckzg==2.0.1
click==8.1.8
coincurve==20.0.0
colorama==0.4.6
crcmod==1.7
cryptography==44.0.0
cytoolz==1.0.1
dataclasses-json==0.6.7
distro==1.9.0
ecdsa==0.19.0
ed25519-blake2b==1.4.1
eth-account==0.13.4
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.1.0
eth-typing==5.1.0
eth-utils==5.2.0
eth_abi==5.2.0
fastapi==0.115.8
filetype==1.2.0
frozenlist==1.5.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
hexbytes==1.3.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.17
langchain-community==0.3.16
langchain-core==0.3.33
langchain-google-genai==2.0.9
langchain-openai==0.2.14
langchain-text-splitters==0.3.5
langgraph==0.2.69
langgraph-checkpoint==2.0.10
langgraph-sdk==0.1.51
langsmith==0.3.3
marshmallow==3.26.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
oauthlib==3.2.2
openai==1.61.0
orjson==3.10.15
packaging==24.2
parsimonious==0.10.0
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
py-sr25519-bindings==0.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
PyJWT==2.10.1
PyNaCl==1.5.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyunormalize==16.0.0
pywin32==308
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rlp==4.0.1
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
standard-imghdr==3.13.0
starlette==0.45.3
tenacity==9.0.0
tiktoken==0.8.0
toolz==1.0.0
tqdm==4.67.1
tweepy==4.15.0
twitter-langchain==0.0.11
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
web3==7.7.0
websockets==13.1
yarl==1.18.3
zstandard==0.23.0
